Updated README and added description on how to start Jupyter

+++ result-analysis/install.sh
@@ -2,11 +2,12 @@
 # Install wireshark and NDN dissector
 echo "Install wireshark and the ndn dissector"
 sudo apt install tshark wireshark
-git clone https://github.com/named-data/ndn-tools.git
-cd ndn-tools/tools/dissect-wireshark
-sudo mkdir /usr/local/share/ndn-dissect-wireshark/
-sudo cp -v ndn.lua /usr/local/share/ndn-dissect-wireshark/
-echo "dofile(\"/usr/local/share/ndn-dissect-wireshark/ndn.lua\")" | sudo tee -a /usr/share/wireshark/init.lua > /dev/null
+if ! test -f /usr/local/share/ndn-dissect-wireshark/ndn.lua; then
+  git clone https://github.com/named-data/ndn-tools.git
+  sudo mkdir -p /usr/local/share/ndn-dissect-wireshark/
+  sudo cp -v ndn-tools/tools/dissect-wireshark/ndn.lua /usr/local/share/ndn-dissect-wireshark/
+  echo "dofile(\"/usr/local/share/ndn-dissect-wireshark/ndn.lua\")" | sudo tee -a /usr/share/wireshark/init.lua > /dev/null
+fi
 
 # Install python dependencies
 echo "Install python virtualenvironment and all dependencies in the icn-analysis environment"
